testmessage: move tests into their own module
